requirements.txt
Flask==3.1.2
flask-cors==6.0.2
gunicorn==23.0.0
pandas==2.3.3
numpy==2.0.2
scipy==1.13.1
openpyxl==3.1.5
xlsxwriter==3.2.9
requests==2.32.5
python-dotenv==1.2.1
rasterio==1.4.3
geopandas==1.0.1
shapely==2.0.7
